BoardService remove refactor v1

+++ service/src/main/java/hu/unideb/inf/rft/neuban/service/impl/BoardServiceImpl.java
@@ -2,6 +2,7 @@ package hu.unideb.inf.rft.neuban.service.impl;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -73,10 +74,27 @@ public class BoardServiceImpl implements BoardService {
 	public void remove(final Long boardId) throws BoardNotFoundException {
 		Assert.notNull(boardId);
 
-		Optional.ofNullable(this.boardRepository.findOne(boardId))
-				.orElseThrow(() -> new BoardNotFoundException(String.valueOf(boardId)));
+		List<UserDto> userDtos = userService.getAll();
 
-		this.boardRepository.delete(boardId);
+		final Optional<BoardEntity> boardEntityOptional = Optional.ofNullable(this.boardRepository.findOne(boardId));
+
+		if (boardEntityOptional.isPresent()) {
+			BoardDto boardDto = modelMapper.map(boardEntityOptional.get(), BoardDto.class);
+			userDtos = userDtos.stream().filter(userDto -> userDto.getBoards().contains(boardDto))
+					.collect(Collectors.toList());
+			if (userDtos == null) {
+				this.boardRepository.delete(boardId);
+			} else {
+				for (UserDto userDtoIter : userDtos) {
+					userDtoIter.getBoards().remove(boardDto);
+					userService.saveOrUpdate(userDtoIter);
+				}
+				this.boardRepository.delete(boardId);
+			}
+		} else {
+			throw new BoardNotFoundException(boardId.toString());
+		}
+
 	}
 
 	@Transactional
